requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cloudinary==1.38.0
distlib==0.3.8
dj-rest-auth==5.0.2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==24.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
filelock==3.13.1
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.0
virtualenv==20.25.0
whitenoise==6.6.0
